fix(eslint): read a hoisted sweep filter in require-table-teardown
